+++ aim/web/ui/src/components/kit_v2/QueryBadge/QueryBadge.style.ts
@@ -1,9 +1,0 @@
-import { styled } from 'config/stitches';
-
-import Text from '../Text';
-
-const ButtonText = styled(Text, {
-  color: '$textPrimary',
-});
-
-export { ButtonText };