+++ aim/web/ui/src/components/kit_v2/Select/SelectItem.tsx
@@ -6,7 +6,7 @@ import { IconCheck } from '@tabler/icons-react';
 import Text from '../Text';
 import Icon from '../Icon';
 import ListItem from '../ListItem';
-import { CheckBox } from '../Checkbox/Checkbox';
+import { Checkbox } from '../Checkbox/Checkbox';
 
 import { ISelectItemProps } from './Select.d';
 
@@ -27,8 +27,13 @@ const SelectItem = ({
   const item = items[index];
   let selected: boolean = false;
   if (item.value) {
-    selected = value === item.value || value?.indexOf(item.value) !== -1;
+    if (value === item.value) {
+      selected = true;
+    } else if (value && Array.isArray(value)) {
+      selected = value.indexOf(item.value) !== -1;
+    }
   }
+
   const rightNode = multiple ? null : selected ? (
     <Icon css={{ color: '$primary100' }} icon={<IconCheck />} />
   ) : null;
@@ -51,7 +56,7 @@ const SelectItem = ({
           key={index}
           size={size}
           rightNode={rightNode}
-          leftNode={multiple ? <CheckBox checked={selected} /> : null}
+          leftNode={multiple ? <Checkbox checked={selected} /> : null}
           onClick={() => onValueChange(item.value!)}
         >
           <Text css={{ color: selected ? '$primary100' : '$textPrimary' }}>