+++ aim/web/ui/src/components/kit_v2/Table/components/TableHeader.tsx
@@ -1,0 +1,19 @@
+import * as React from 'react';
+import classNames from 'classnames';
+
+import { TableHeaderStyled } from '../Table.style';
+
+const TableHeader = React.forwardRef<
+  HTMLTableSectionElement,
+  React.HTMLAttributes<HTMLTableSectionElement>
+  /* eslint-disable react/prop-types */
+>(({ className, ...props }, ref) => (
+  <TableHeaderStyled
+    ref={ref}
+    className={classNames('TableHeader', className)}
+    {...props}
+  />
+));
+
+TableHeader.displayName = 'TableHeader';
+export default TableHeader;