+++ CHANGELOG.md
@@ -1,16 +1,19 @@
 # Changelog
+
 ## Unreleased
 
-### Enhancements
+### Enhancements:
+- Add ability to create reports (mihran113)
 - Add support for self-signed SSL certificates (mihran113)
+
 
 ## 3.24.0 Aug 14, 2024
 
-### Enhancements
+### Enhancements:
 - Add read-only mode for Aim UI (mihran113)
 - Support of mass updates in remote tracking (peter-sk)
 
-### Fixes
+### Fixes:
 - Fix bug in bookmark page where it was not scrollable if there was too many bookmarks (vinayan3)
 - Fix exception name in `storage/union.pyx` (sulan)